+++ example/metro.config.js
@@ -1,9 +1,12 @@
 const path = require('path');
-const blacklist = require('metro-config/src/defaults/blacklist');
+const { getDefaultConfig } = require('metro-config');
+const exclusionList = require('metro-config/src/defaults/exclusionList');
 const escape = require('escape-string-regexp');
 const pak = require('../package.json');
 
 const root = path.resolve(__dirname, '..');
+
+const { resolver: defaultResolver } = getDefaultConfig.getDefaultValues();
 
 const modules = Object.keys({
   ...pak.peerDependencies,
@@ -14,9 +17,9 @@ module.exports = {
   watchFolders: [root],
 
   // We need to make sure that only one version is loaded for peerDependencies
-  // So we blacklist them at the root, and alias them to the versions in example's node_modules
+  // So we exclude them at the root, and alias them to the versions in example's node_modules
   resolver: {
-    blacklistRE: blacklist(
+    blacklistRE: exclusionList(
       modules.map(
         (m) =>
           new RegExp(`^${escape(path.join(root, 'node_modules', m))}\\/.*$`)
@@ -27,6 +30,8 @@ module.exports = {
       acc[name] = path.join(__dirname, 'node_modules', name);
       return acc;
     }, {}),
+
+    sourceExts: [...defaultResolver.sourceExts, 'cjs'],
   },
 
   transformer: {