+++ src/__tests__/Logger.spec.ts
@@ -1,4 +1,5 @@
 import XHRInterceptor from 'react-native/Libraries/Network/XHRInterceptor';
+import { warn } from '../utils/logger';
 import Logger from '../Logger';
 
 jest.mock('react-native/Libraries/Blob/FileReader', () => ({}));
@@ -12,12 +13,19 @@ jest.mock('react-native/Libraries/Network/XHRInterceptor', () => ({
   enableInterception: jest.fn(),
 }));
 
+jest.mock('../utils/logger', () => ({
+  warn: jest.fn(() => {
+    throw new Error('Unexpected warning');
+  }),
+}));
+
 beforeEach(() => {
   jest.clearAllMocks();
 });
 
 describe('enableXHRInterception', () => {
   it('should do nothing if interceptor has already been enabled', () => {
+    (warn as jest.Mock).mockImplementationOnce(() => {});
     const logger = new Logger();
 
     (XHRInterceptor.isInterceptorEnabled as jest.Mock).mockReturnValueOnce(
@@ -25,8 +33,21 @@ describe('enableXHRInterception', () => {
     );
 
     expect(logger.enableXHRInterception()).toBeUndefined();
+    expect(warn).toHaveBeenCalledTimes(1);
     expect(XHRInterceptor.isInterceptorEnabled).toHaveBeenCalledTimes(1);
     expect(XHRInterceptor.setOpenCallback).toHaveBeenCalledTimes(0);
+  });
+
+  it('should continue if interceptor has already been enabled but forceEnable is true', () => {
+    const logger = new Logger();
+
+    (XHRInterceptor.isInterceptorEnabled as jest.Mock).mockReturnValueOnce(
+      true
+    );
+
+    expect(logger.enableXHRInterception({ forceEnable: true })).toBeUndefined();
+    expect(XHRInterceptor.isInterceptorEnabled).toHaveBeenCalledTimes(1);
+    expect(XHRInterceptor.setOpenCallback).toHaveBeenCalledTimes(1);
   });
 
   it('should update the maxRequests if provided', () => {
@@ -37,6 +58,17 @@ describe('enableXHRInterception', () => {
     const maxRequests = logger.maxRequests;
 
     expect(maxRequests).toBe(23);
+  });
+
+  it('should update the ignoredHosts if provided', () => {
+    const logger = new Logger();
+
+    logger.enableXHRInterception({ ignoredHosts: ['test.example.com'] });
+    // @ts-ignore
+    const ignoredHosts = logger.ignoredHosts;
+
+    expect(ignoredHosts).toBeInstanceOf(Set);
+    expect(ignoredHosts?.has('test.example.com')).toBeTruthy();
   });
 
   it('should call setOpenCallback', () => {
@@ -158,5 +190,39 @@ describe('getRequest', () => {
     // @ts-expect-error
     const result = logger.getRequest(2);
     expect(result).toBe(requests[0]);
+  });
+});
+
+describe('openCallback', () => {
+  it('should add new request to start of requests list', () => {
+    const logger = new Logger();
+    logger.enableXHRInterception();
+
+    const xhr = {};
+    const url1 = 'http://example.com/1';
+    const url2 = 'http://example.com/2';
+
+    // @ts-expect-error
+    logger.openCallback('POST', url1, xhr);
+    // @ts-expect-error
+    logger.openCallback('POST', url2, xhr);
+    expect(logger.getRequests()[0].url).toEqual(url2);
+    expect(logger.getRequests()[1].url).toEqual(url1);
+  });
+
+  it('should ignore requests that have ignored hosts', () => {
+    const logger = new Logger();
+    logger.enableXHRInterception({ ignoredHosts: ['ignored.com'] });
+
+    const xhr = {};
+    const url1 = 'http://example.com/2';
+    const url2 = 'http://ignored.com/1';
+
+    // @ts-expect-error
+    logger.openCallback('POST', url1, xhr);
+    // @ts-expect-error
+    logger.openCallback('POST', url2, xhr);
+    expect(logger.getRequests()[0].url).toEqual(url1);
+    expect(logger.getRequests()).toHaveLength(1);
   });
 });